created a new file EventContextMenu and added a isDoubleClick useState to prevent an error message occurring on delete
